requirements.txt
affine==2.0.0.post1
appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
chainer==1.22.0
cupy==1.0.0a1
cycler==0.10.0
dask==0.14.0
decorator==4.0.11
filelock==2.0.7
h5py==2.6.0
Keras==1.2.2
matplotlib==2.0.0
mecab-python3==0.7
msgpack-numpy==0.3.9
msgpack-python==0.4.8
networkx==1.11
nose==1.3.7
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
Pillow==4.0.0
plyvel==0.9
protobuf==3.1.0
py==1.4.32
pycuda==2016.1.2
pyparsing==2.2.0
pytest==3.0.7
python-dateutil==2.6.0
pytools==2016.2.6
pytz==2016.10
PyYAML==3.12
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==0.12.0
Theano==0.8.2
toolz==0.8.2
Werkzeug==0.12.1
